+++ base/NCIO.F90
@@ -26,10 +26,11 @@ module mapl_NCIO_mod
   use mapl_EASEConversion_mod, only: MAPL_get_ease_gridname_by_cols => get_ease_gridname_by_cols
 
   use mapl_Shmem_mod
-  use mapl_ExceptionHandling_mod
+  use mapl_ErrorHandling_mod
   use netcdf
   use pFIO
   use MAPL_Constants
+  use mapl_ErrorHandling_mod
   use gFTL2_StringIntegerMap
   use gFTL2_StringVector
   use, intrinsic :: ISO_C_BINDING