+++ pfio/ReplaceMetadataMessage.F90
@@ -2,13 +2,14 @@
 #include "unused_dummy.H"
 
 module pFIO_ReplaceMetadataMessageMod
-   use mapl_ExceptionHandling_mod
+   use mapl_ErrorHandling_mod
    use pFIO_UtilitiesMod
    use pFIO_FileMetadataMod
    use pFIO_AbstractMessageMod
    use pFIO_StringVariableMapMod
    use pFIO_StringVariableMapUtilMod
    use mapl_KeywordEnforcer_mod
+   use mapl_ErrorHandling_mod
    implicit none
    private