+++ pfio/ServerThread.F90
@@ -7,7 +7,7 @@ module pFIO_ServerThreadMod
    use, intrinsic :: iso_c_binding, only: c_loc
    use, intrinsic :: iso_fortran_env, only: REAL32, REAL64, INT32, INT64
    use, intrinsic :: iso_c_binding, only: c_f_pointer
-   use mapl_ExceptionHandling_mod
+   use mapl_ErrorHandling_mod
    use mapl_Profiler_mod
    use pFIO_UtilitiesMod, only: word_size, i_to_string
    use pFIO_AbstractSocketMod