+++ pfio/UnlimitedEntity.F90
@@ -26,7 +26,7 @@ module pFIO_UnlimitedEntityMod
 
    use pFIO_ConstantsMod
    use pFIO_UtilitiesMod
-   use mapl_ExceptionHandling_mod
+   use mapl_ErrorHandling_mod
    use, intrinsic :: iso_fortran_env, only: INT32, INT64
    use, intrinsic :: iso_fortran_env, only: REAL32, REAL64
 
@@ -726,7 +726,7 @@ module pFIO_StringUnlimitedEntityMapUtilMod
    use pFIO_UtilitiesMod
    use pFIO_UnlimitedEntityMod
    use pFIO_StringUnlimitedEntityMapMod
-   use mapl_ExceptionHandling_mod
+   use mapl_ErrorHandling_mod
    implicit none
    private