+++ utils/API.F90
@@ -16,9 +16,6 @@ module mapl_utils_api
    public :: MAPL_Return
    public :: MAPL_Deprecated
    public :: MAPL_SetFailOnDeprecated
-   public :: MAPL_RTRN
-   public :: MAPL_Vrfy
-   public :: MAPL_ASRT
    public :: MAPL_abort
    public :: MAPL_set_abort_handler
    public :: MAPL_SUCCESS