+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed old interfaces MAPL_RTRN, MAPL_Vrfy, MAPL_ASRT, mapl_ExceptionHandling_mod
+
+### Changed
+
 - Aligned API umbrellas with actual content.  Earlier refactoring had
   done some weird indirect publishing.
 - Eliminate some items from public interface of MAPL.